Refactored component structure
