+++ src/components/BookMeeting/ReportIssue/Partials/ReportIssueButton.jsx
@@ -1,22 +1,8 @@
-import { useState } from 'react';
+// MUI Imports
 import { Typography, Button } from '@mui/material';
-
 import NotificationsNoneOutlinedIcon from '@mui/icons-material/NotificationsNoneOutlined';
-import ReportIssuePopup from './Popup/Main';
 
-const ReportIssue = () => {
-  const [open, setOpen] = useState(false);
-  const handleClickOpen = () => setOpen(true);
-  const handleClose = () => setOpen(false);
-
-  return (
-    <>
-      <ReportIssueButton handleClickOpen={handleClickOpen} />
-      <ReportIssuePopup open={open} handleClose={handleClose} />
-    </>
-  );
-};
-
+// Button that triggers the popup for Reporting an Issue
 const ReportIssueButton = ({ handleClickOpen }) => {
   return (
     <Button
@@ -38,4 +24,4 @@ const ReportIssueButton = ({ handleClickOpen }) => {
   );
 };
 
-export default ReportIssue;
+export default ReportIssueButton;
