+++ src/components/BookMeeting/ReportIssue/Popup/Main.jsx
@@ -1,8 +1,13 @@
+// MUI Imports
 import { Dialog, DialogContent } from '@mui/material';
+
+//Local Imports
 import PopupTitle from './PopupTitle';
 import PopupButtons from './PopupButtons';
 import PopupFormFields from './PopupFormFields';
 
+// This is the main component, creates a popup with a form
+// for reporting issues with the app (or smth. else)
 const ReportIssuePopup = ({ open, handleClose }) => {
   return (
     <Dialog open={open} onClose={handleClose}>
