+++ src/components/BookMeeting/ReportIssue/Popup/PopupButtons.jsx
@@ -1,5 +1,8 @@
+// MUI Imports
 import { DialogActions, Button, Typography } from '@mui/material';
 
+// Buttons for closing the popup
+// P.S. The report button should eventually process the form data and send it to the server
 const PopupButtons = ({ handleClose }) => {
   return (
     <DialogActions>
